requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
discord.py==2.3.2
frozenlist==1.4.1
greenlet==0.4.13
hpy==0.0.4.dev179+g9b5d200
idna==3.6
multidict==6.0.5
psycopg2cffi==2.9.0
pycparser==2.21
PyNaCl==1.5.0
PyYAML==6.0.1
readline==6.2.4.1
requests==2.31.0
six==1.16.0
urllib3==2.2.0
yarl==1.9.4
